requirements.txt
pyTelegramBotAPI
flask
openai
python-dotenv
gunicorn